Handle NoAddrsAvail/NoPrefixAvail errors during Solicits.
According to RFC3315, Sec 17.2.2, if a "server will not assign any
addresses ... it MUST send an Advertise message ... with code
NoAddrsAvail". RFC3633 Sec 11.1 specifies the same about
"NoPrefixAvail" errors.

Hence, if a client specifies *both* an IA_NA *and* an IA_PD option
in a Solicit request from a server that is configured to only handle
prefix delegations, or vice versa, that server will return an error.

Previously, in that case the client would get stuck retrying the
exact same Solicit request. However, if the user specified that
either address assignments or prefix delegations are optional, then
the client can instead retry the Solicit request just an IA_NA or
IA_PD option (instead of both). That way it would still be able to
successfully receive a prefix delegation from a server that does not
serve address assignments, or an address assignment from a server
that does not serve prefix delegations.

Additionally, if a client receives an UnspecFail error, it should
not try to handle the response, and instead retry later. RFC7083 Sec
7 specifies that:

    If a DHCPv6 client receives a message containing a SOL_MAX_RT
    option that has a valid value for SOL_MAX_RT, the client MUST
    set its internal SOL_MAX_RT parameter

So I've assumed that SOL_MAX_RT/INF_MAX_RT options should still be
handled even in the case of UnspecFail errors.