odhcp6c: fix deamon raw buffer inc
When starting odhcp6c as a daemon using the following command:
  odhcp6c -s /etc/odhcp6c.script -p /var/run/eth0_odhcp6c.pid -P 0 \
    -t 120 eth0"
The DHCPv6 server sends RA packets every 10 seconds.
However, it was observed that the raw socket's receive buffer keeps
increasing continuously. After investigating the code, an error was found:
the I/O signal was bound before starting the daemon, causing the daemon to
not receive the I/O signal. Consequently, the raw socket's recv queue kept
growing indefinitely.

The author of this patch is inactive and didn't include a proper SoB:
https://github.com/Horbivores

Link: https://github.com/openwrt/odhcp6c/pull/80
Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>