Remove access to uninitialised array
Remove an access to an uninitialised array, which is causing error
handling to be inadvertently skipped for replies containing IA_PD or
IA_NA with an error status code.

`handled_status_codes` is an array of length 7, but only the first
element is initialised.  This array looks like it was intended as a
mechanism for indicating which status codes should be considered errors,
but this mechanism is not actually used.