odhcp6c: do not send RELEASE only if -k is set and the process is stopped
Problem: Based on the README, if -k option is set, RELEASE message should not be sent when the process is stopped.
However RELEASE message are also not sent in case of a forced RELEASE or reconfiguration.

Solution: Stop the transmission of RELEASE message whene the process is stopped but allow it for a forced RELEASE or reconfiguration.