dhcpv6: always include both IA_NA and IA_PD in Request message if requested
The RFC7550 says that if you request both IA_NA and IA_PD and only one option is offered in the Advertise,
then the client must send a Request message (and not a Solicit) and the message SHOULD still include both IA_NA and IA_PD options.

Problem: A Request message is correctly send instead of a Solicit but it doesn't include both IA_NA and IA_PD options.

Solution: If IA_NA and IA_PD are requested, the Request message always include both options.