dhcpv6: state handling bugfix
It appears that somewhere between 96017df54d8fd6ec1 and
63461f64d4c162dc2 the state handling changed and we no longer send
an INFORMATION-REQUEST that would otherwise provide DNS (etc)
servers. When in SLAAC only configuration, we should still do this.

This mode is common with some ISP in Japan and likely elsewhere.

This has been tested and verified to resolve the problem noted in
issue #134. While each section separately resolves the issue, both
are included.

Closes #134

Signed-off-by: Paul Donald <newtwen+github@gmail.com>